requirements_dev.txt
django-debug-toolbar==1.9.1
django-extensions==1.9.8
